+++ braid/Changelog.md
@@ -22,6 +22,8 @@
     + Note that index errors will be different for names. e.g. instead of 'out
         of bounds', the error will be 'name does not exists'
 - Some error variants have changed due to indexing changes
+- `OracleT::rowsim` and `OracleT::rowsim_pw` use `RowSimiliarityVariant` instead
+    of the boolean `col_weighted` argument.
 
 ## 0.39.5
 - Minor optimization in joint entropy computation between one categorical and